pyproject.toml
[project]
name = "redaction-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "dotenv>=0.9.9",
    "easyocr>=1.7.2",
    "fastapi>=0.116.1",
    "gliner[tokenizers]>=0.2.21",
    "huggingface-hub>=0.33.4",
    "ipdb>=0.13.13",
    "onnxruntime>=1.22.1",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.9",
    "pymupdf>=1.23.5",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "torch>=2.7.1",
    "transformers>=4.53.2",
    "uvicorn>=0.35.0",
]

[tool.setuptools]
packages = ["app", "alembic"]
